[+BUGFIX] Extbase (Object): Refactor Object Manager
The Object Manager is now at the same location and
has the same API as in FLOW3.

[+BUGFIX] Extbase: Major cleanups to Dependency Injection and Persistence

Now, DI finally works with Persistence, cleaning
this greatly up. Additionally, all internal
t3lib_div::makeInstance calls have been replaced.
Now, dependency injection is actually usable.

Additionally, we completely thought over which
persistence classes need to be singleton and which
should be prototype, leading finally to a
coherent design in the persistence layer.

[+BUGFIX] Extbase: remove non-used interfaces

Removed classes which were not used.

Relates to: #9062
Resolves: #10585
Resolves: #10564