[+FEATURE] Extbase (Utility): Add support for custom specify custom plugin icon
Tx_Extbase_Utility_Extension::registerPlugin() now has a fourth parameter $pluginIconPathAndFilename
that you can set to the path of an icon, that will be displayed in the plugin dropdown in the backend.
You can use it like:
Tx_Extbase_Utility_Extension::registerPlugin(
	$_EXTKEY,
	'MyPlugin',
	'My Plugin description',
	t3lib_extMgm::extRelPath($_EXTKEY) . 'Resources/Public/Icons/someIcon.gif'
);

Thanks to Steffen Müller for the initial patches!

Resolves #11179