[!!!][+BUGFIX] make backend-modules in web use new pagetree by default
In TYPO3 4.5 the new pagetree was introduced. It's possible to use
this in an Extbase backend-module with the config option
'navigationComponentId' in Tx_Extbase_Utility_Extension::registerModule.

With introducing this as default config for modules in 'web', we make
all Extbase based backend-modules make use of the new pagetree.

Even though this is a breaking change, all backend-modules are asked
to use the new pagetree since TYPO3 4.5 anyway.

* make navigationComponentId=typo3-pagetree default config

Change-Id: I3f08134a58bf41c7940ac3050f6b4843988a30b4