[+BUGFIX] Revert incorrect deletion of Query::count() & Query::withUid()
The methods Tx_Extbase_Persistence_Query::count() and
Tx_Extbase_Persistence_Query::withUid() are deprecated.
They have been removed in
Ic1b6217bc6c9100b6c48c5d30eca6068417265a8.

But the deletion is incomplete and leads to a fatal error
We'll have to keep count() until version 1.5 according to the
TYPO3 release policy.

WithUid() is already deprecated since extbase 1.2 so it could
be removed with 1.4. But it has to be removed from the interface
and from all parts in the core that are still using it (currently two)

Relates: #13745

Change-Id: I390edfe3de1fe92a27723dcccdd6a6d098f5ac07