[FEATURE] Improve information for detail view
The row details view in the Workspaces module is based on two arrays
containing only the changed fields, one with the Live value of the
fields and one with the diff'ed values. These arrays do not contain
any reference to the name of the fields, only to their labels.

Add an extra entry for each difference, containing the field name.

Change-Id: Ie2b4088a568a4f9880afdf44e3b5f95cf95215fc
Resolves: #27442
Releases: 4.6, 4.5