[!!!][+FEATURE] (Validation): Adjusted and re-done Validation API
* Completely reworked Validation API, adjusting all validators.
  Validators now return a Error_Result object.
* Validators can now be singletons or prototypes
* Validator options are passed in the constructor

The whole change is *backwards-compatible*, so all default validators can be
used with the new MVC stack in Extbase 1.4, and with the MVC stack in
compatibility mode.

The same goes for *custom validators* which extend *AbstractValidator*. They
are automatically upwards-compatible.

!!! If an extension author directly implemented the *ValidatorInterface*,
he needs to adjust his validator to support the new API.

Resolves: #27653
Change-Id: I6b3c9f4a629728f7e1fa2cd672f38f69c7321f52