[+FEATURE] (Configuration): Feature flag implementation
In order to provide new features and stay backwards-compatible at the same time,
the framework has to distinguish at some points between old code and new code
with breaking changes.
At these points, the feature flag should be used.

Note: This is right now only an INTERNAL API, not to be used by Extension
Developers yet. It is needed as a basis for the new Property Mapper and MVC
adjustments, but it this solution will be revised at the T3DD11.

Related: #27070
Related: #27059
Change-Id: I29d79efb359ae8e61a8e4c08cd04037296e33720