[TASK] Timestamp support for DateTimeConverter
This change adds DateTimeConverter support for integer values
by using the "U" dateFormat which reads a UnixTimestamp date
format (i.e. integer as string).

Under the assumption that the new PropertyMapper is being
used, this fixes the bugs below.

At the same time backport the changes done in Flow and use the
same solution which is currently pending in Flow (#41989)

Fixes: #7574
Fixes: #8536
Fixes: #5595
Releases: 6.0

Change-Id: Id39e37451767ea732288423911109d83d47aa6e4
Reviewed-on: http://review.typo3.org/4440
Reviewed-by: Tymoteusz Motylewski
Tested-by: Tymoteusz Motylewski