[FEATURE] Include preview link in notification
Until now there haven't been a preview link in the notification mails.

With this feature there are two new markers that can be used in the
notification template. One marker for the workspace preview link and
one marker for the splitted workspace preview link.
Additionally this feature adds a hook, so it is possible for a
developer to add severall additional markers.

Belongs to change 1054
Resolves: #13518

Change-Id: I8443718f61dcdb066a8935f424ee6bb1a86bd0fc
Reviewed-on: http://review.typo3.org/1057
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader