[BUGFIX] Revert change in interface signature
The type hints in the interface signature break other extensions adding
linkvalidator tasks. This is because they extend the abstract linktype
class and the interface was without type hints (array, object name) in
4.5.

Thus it is impossible to support both 4.5 and 4.6 with the same
linktype, although the interface is the very same.

Solution: Drop the hints again (PHPdoc must be enough).

Change-Id: I3c344a89344825e3ce55d46ec03f431287c9bf8b
Fixes: #30589
Related: #M18276