[BUGFIX] With "virtual page-root" review-grid needs restriction  to mountpoints visible
With the "virtual page-root"-feature (#33852) a person which is
not admin now has a root-node to click on to get a full view of
changes across all mountpoints (requested in #30630).

If person may not see the "real" page-root (pid=0) the selection
needs to be limited to changes from all visible mountpoints.

Change-Id: I7be7191d6095005c18f05067f1350e0ccab3e67e
Fixes: #34308
Releases: 4.7
Reviewed-on: http://review.typo3.org/9248
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Philipp Gampe
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter