[BUGFIX] Custom validators may be overriden by default validators
The softref parser always puts the type to 'db' for db records.
Due to the abstract validators fetchType method just returning this
type, the type gets overriden if any validator without own fetchType
method is evaluated later.
The solution is to store the type inside the record array, so that the
returned type is the last special type found.

Change-Id: I5d831c4f25c37da2606ad772e91bcfa11891e1df
Fixes: #34566
Releases: 4.5,4.6,4.7,6.0