[TASK] Improve method makeCategorizable to be used everywhere
Method makeCategorizable registers a table to be categorizable in the
BE. The current implementation has the drawback to force the call of this
method after the TCA of the table. If called before, the "categorization"
wouldn't be applied because of timing issue. The patch improve the
situation by postponing the application of the TCA responsible for
categorization.

For more details about the method refer to
\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::makeCategorizable.

Change-Id: I966cc65e50e0d6a81cfafb1238f253ed802349c6
Resolves: #38972
Releases: 6.2, 6.1
Reviewed-on: https://review.typo3.org/12812
Reviewed-by: Francois Suter
Tested-by: Francois Suter
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe