[BUGFIX] PHP 5.4 unit test warnings
PHP 5.4 throws warnings "Creating default object from empty value"
if $foo->bar is accessed and $foo is not an object. This is the
case in several unit tests, especially if $GLOBALS['TSFE'] is accessed.
The patch fixes those warnings by initializing the object in question
with an instance of stdClass.
Furtermore some dumplicate cleanups of global backup handling is
removed that is already done in the parent classes.

Change-Id: I4c060632783926d76ff96b2eadf3c2ae89f6ae4e
Fixes: #42610
Releases: 6.0
Reviewed-on: http://review.typo3.org/16116
Reviewed-by: Helmut Hummel
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn