[TASK] Rename property, getter and setters of File Model
To be in line with TYPO3 Flow, rename the properties from
non speaking "object" to "originalResource"

This change did not make it into 6.0 final, but still should
be part of the 6.0 branch. Since the file models were not
usable anyway because of a bug in the data mapper
(see #10770), it's safe to rename these.

Resolves: #44255
Releases: 6.0, 6.1

Change-Id: Id9efe02a80e29e1ac6a4f6697f104bfb59e51399
Reviewed-on: https://review.typo3.org/17287
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Reviewed-by: Mattias Nilsson
Reviewed-by: Stefan Neufeind
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel