[!!!][BUGFIX] Remove support for TCA setting 'foreign_selector'
The TCA setting 'foreign_selector' is a UI feature which can be
configured for field in a relation table, when this table
is editable using the IRRE feature. Then a selector box
is rendered making it possible to select records from a child
table.

In this scenario the relation table itself may contain other
editable fields, which should be represented by a domain model.

The latter is currently not possible with Extbase because
'foreign_selector' triggers the DataMapper to resolve a
many to many relation to the child table.

Solution is to remove this behaviour and let the DataMapper
resolve the one to many relation to the relation table.

Find a test setup at https://gist.github.com/4284545

Change-Id: I2d8366f9ce54470999c1b02ad2098359f5aaa6b3
Fixes: #10770
Releases: 6.0, 6.1
Reviewed-on: https://review.typo3.org/17159
Reviewed-by: Thomas Maroschik
Reviewed-by: Sebastian Michaelsen
Tested-by: Sebastian Michaelsen
Reviewed-by: Marc Bastian Heinrichs
Tested-by: Marc Bastian Heinrichs
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Arjen Hoekema
Tested-by: Arjen Hoekema
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring