[BUGFIX] Media element not working with FAL
With FAL the path to a file is no longer used,
instead the UID to the specific file is used,
prefixed with "file:"

This patch just adds a check if the filename
starts with "file:" and if so load the specific FileObject
to retrive the url. This check is done before all others,
because FAL should be the default file handler.
Additionally use explode() first to get the filename
if an editor adds (non used)
link parameters: file:740 _blank media-title

This fixes also the problem with audio files mentioned in #39416,
using the same patch.

Resolves: #41280
Related: #39416
Releases: 6.0, 6.1
Change-Id: Ibec6cefda54b6e136c0d5ae43c67ef2eea3b1bb8
Reviewed-on: https://review.typo3.org/18228
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Marco Huber
Tested-by: Marco Huber
Reviewed-by: Stefan Neufeind
Reviewed-by: Mattias Nilsson
Tested-by: Mattias Nilsson
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn