[BUGFIX] Skip unreliable APC test on PHP versions below 5.3.4
APC has some slam protection that tries to prevent hammering of cache
entries. This can be disabled, but the option does not work at least
in native PHP 5.3.3 on debian squeeze, While it is no problem with
higher PHP version like the current one on travis-ci.org,
the test is now just skipped on PHP environments that are knows for issues.