[!!!][-FEATURE] Remove ext:statictemplates
ext:statictemplates is unmaintained for a long time, the
functionality isn't used anymore in most projects, the
code and templates are extremely old school.

The patch throws the extension away and adapts some
files that referenced things from it.

The following MENU objects are broken now, they are
very likely unused in most projects, too. They should
work again if the extension is loaded again.
* GMENU_LAYERS
* TMENU_LAYERS
* GMENU_FOLDOUT