[TASK] Update file type according to Iana standard
This change relates to the File types in FAL.
The Iana standard http://www.iana.org/assignments/media-types
defines a list of Media types which are followed in TYPO3 CMS.
It looks the "Software" type should be deprecated and replaced by
"Application". This matches the mime-type e.g application/pdf and
sounds more comprehensive.

Change-Id: Ic2ec7173d10b6139e6e2a19fcead3086665a3aae
Fixes: #45503
Releases: 6.1
Reviewed-on: https://review.typo3.org/18297
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack