[!!!][TASK] Remove t3lib_formmail
This class should have been deprecated for 6.0 but was forgotten.
The formmail API is heavily old school and hard coded. It is of
no use in younger projects for a long time.
The class is removed now without substitution and the according
helpermethods in TSFE deprecated, checkDataSubmission() even removed
directly because it can not be used externally.