[TASK] Move t3lib/fonts to core/Resources/Private/Font
vera.ttf is only used in the install tool in 'gd library functions'
to test for imagemagick / graphicsmagick compatibility.
nimbus.ttf is used as fallback font in GIFBUILDER if no different
font is given.
Both files are moved to core/Resources/Private/Font, with adapted
code mentions. The source directory is removed, there is no need
to deploy them according to the license of those files.