[TASK] Second TCA cache layer for frontend rendering
The patch introduces a second cache layer used during frontend
rendering. On first access, full ext_tables is executed and the
result TCA is cached for next access.
This way, TCA is always available in frontend, so extensions like
extbase which need TCA for mapping stuff or similar extensions
do not need to include ext_tables on their own.