[TASK] Move several system checks to new Check class
* checkMaximumFileUploadSize
* checkPostUploadSizeIsHigherOrEqualMaximumFileUploadSize
* checkMemorySettings
* checkPhpVersion

The implemented checks for memory settings and php version make the
ext:install/requirements.php file obsolete. The two tests are also
duplicated by the reports module and are removed for now. The idea
is, that the reports module gets an adapter to the new environment
check class later on to show *all* messages.
The check for minimum php version can be left out this way, too.