[TASK][Check] Add check for PHP opcode cache and doc comment reflection
A general test with warning severity is added if neither apc, xcache nor
eaccelerator is loaded.
The checks for working doc comment reflection from install tool
and a reports test in extbase are moved to a check method. This fails if
eaccelerator is not compiled with --with-eaccelerator-doc-comment-inclusion
flag and raises an error status now.