[BUGFIX] Warning empty $GLOBALS['TCA'] in rootline utility
The TCA refactoring of issue #45767 introduced a warning in the
rootline utility: Rootline resolving is called by determineID in
index_ts. At this point in bootstrap the TCA was not loaded yet.
The patch just moves the TCA includes prior to determineID.

Warning: This leads to an earlier execution of ext_tables.php if
caches are empty. 6.0 did that too, it called includeTca()
in RootlineUtility to circumvent the same issue, the current patch
just solves the issue in a cleaner way. But still, this might lead
to problems if ext_tables.php code relies for example on the setup
done by determineID or the other code in between. The risk this
breaks something is pretty low, though.

Change-Id: I83e6367dbb4654fe0bf957bbe817fa6d5b5ea106
Resolves: #46156
Related: #45767
Releases: 6.1
Reviewed-on: https://review.typo3.org/19200
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn