[BUGFIX] Folder Tree does not respect _temp_ and _recycler_
The Folder Tree View does not respect temp and
recycler settings, which are identified by the folder
names.

The file list itself does that, because there is a wrong
use of "getIdentifier".

You can see that the foldertree does not show
"Temporary" but "_temp_"  because the comparison
fails every time. The file list works however.

Resolves: #46603
Releases: 6.1, 6.0
Change-Id: I08de29b737ef3464addceda9f610e9dd0a7b1034
Reviewed-on: https://review.typo3.org/19222
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Felix Nagel
Tested-by: Felix Nagel
Reviewed-by: Tom Ruether
Reviewed-by: Andreas Wolf
Tested-by: Andreas Wolf