[BUGFIX] Fix persisting of lazy object storages
Lazy object storages saved in the clean state of a
parent object get initialized during persisting and
set the clean state of the parent object to the current changed
state. After that the persistence backend cannot detect any
changes and skips persisting of the changed relations.

This fix checks if the lazy object storage is already part
of the parents clean state and skips the memorization in
this case.

Fixes: #47703
Releases: 6.2, 6.1
Change-Id: I6801ca8ff709253ac33c65bc619b5d794f5615ce
Reviewed-on: https://review.typo3.org/20335
Reviewed-by: Nico de Haen
Tested-by: Nico de Haen
Reviewed-by: Marc Bastian Heinrichs
Tested-by: Marc Bastian Heinrichs