[FEATURE] Backport Object Type Converter from Flow
In Flow there is a Type Converter which can map array sources
to not persistent objects. This is very useful if you need
transitional objects built from request arguments.

Backporting this converter needs some minor modifications
in the reflection service, which are included in this commit
as well as registering the new converter.

Resolves: #48548
Releases: 6.2

Change-Id: Ic88b732076ae19ece490cf1376b2d1bbcaf1ebff
Reviewed-on: https://review.typo3.org/21114
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Daniel Hürtgen
Tested-by: Daniel Hürtgen
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel