[BUGFIX] be_users.userMods is too small to hold all modules
It is impossible to assign all modules to a single user, because the
DB field is too small.  Saving such a record will result in a DB error
and a truncated modules list.

Solution: Change to field from varchar(255) to text to support a larger
number of modules.

This change requires a DB compare to become active. The fix is simply
not present otherwise, but also don't affect existing installations in
any way.

Resolves: #16762
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I67b2ed6b650f9eaf5bdcc9ef81ce93cbbdc54b7f
Reviewed-on: https://review.typo3.org/20885
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny