[FEATURE] Install tool: Check xdebug.max_nesting_level
If PHP xdebug extension is loaded, TYPO3 CMS can fail with the
xdebug.max_nesting_level default value of 100. This is easily
triggered by fluid, and since we have fluid in the install tool
we need a check that gives an administrator a nice warning if the
nesting level is not set high enough for our demands.

Change-Id: I4812e7f47b04219d380d7454d51b2e6b74436b6d
Resolves: #49298
Related: #49162
Releases: 6.2
Reviewed-on: https://review.typo3.org/21549
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn