[BUGFIX] Undefined switchable controller actions must be removed
Currently there is no check if switchable controller actions in
pi_flexform are still present in the configuration flexform. So,
when removing the flexforms in the configuration file, the
switchable controller actions section is not deleted in pi_flexform.

This change removes all occurences in the present pi_flexform, so
if no switchable controller actions are defined any longer, they
will be removed, else they are overwritten by the current flexform
data.

Releases: 6.2
Resolves: #31198
Change-Id: I6046b29dbf436d8c970593dbc6fd2c216f222c52
Reviewed-on: https://review.typo3.org/22030
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring