[BUGFIX] Workspaces and Version update wizard isn't skippable
When you skip the update wizard for workspaces and version
you always get the wizard back. The update wizard is not marked
as done when you don't want to install the extensions.

To fix this we need to copy some functions from AbstractUpdate
because the update wizard for workspaces and version extend the
SystemExtension update wizard.

Change-Id: I4f7f3e8606cdf41db5737b5f4dd22d8faf4a4acb
Resolves: #47734
Releases: 6.2
Reviewed-on: https://review.typo3.org/22680
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring