[!!!][TASK] Clean up OpendocsToolbarItem references
With toolbar refactoring, OpendocsController was renamed
to OpendocsToolbarItem and a class alias was introduced to
cope with core internal usages of the old class names. This
is cleaned up with this patch.

The class was still referenced in alt_doc_nodoc.php entry
point through EditDocumentController. This ancient code is
obsolete since around 4.3 but was never removed. The patch
now removes this entry point and redirects to dummy.php in
EditDocumentController just in case this code snippet is
still triggered somehow - which is unlikely, but can not be
overseen currently.

Furthermore, the alias OpendocsController is removed: The
new class implements a different interface now, so this
would break anyway - which is also unlikely to be used by
third party code.

Change-Id: Iaa5a7be7ef0e50c1ca1aca245c163e7bc5790410
Resolves: #63110
Related: #62928
Releases: master
Reviewed-on: http://review.typo3.org/34409
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>