[BUGFIX] Clean up "core" documentation
Clean all reStructuredText markup errors that generate
warnings during the rendering of the "core" sysext
documentation.

Furthermore a link is added to the "Howto" file, so that it
is easier to find the instructions in the rendered version.

Finally a Settings.yml file is added, just like all other
manuals, with a title and version number.

Resolves: #64431
Releases: master
Change-Id: Ic75cad4d1d364aceef04a06f96b7fd74d454559c
Reviewed-on: http://review.typo3.org/36174
Reviewed-by: Francois Suter <francois@typo3.org>
Tested-by: Francois Suter <francois@typo3.org>