[FormEngine] Introduce option array from FormEngine to childs
Introduces an option array given from FormEngine to child elements.
AbstractFormElement is extended to get those options. Option
renderReadonly is adapted to use this and the recently introduced
setRenderReadonly() is removed.

This option array will be extended to additional options that
need to be given from FormEngine to childs later.

Change-Id: I480826dfe2e59d3a6559f9bb9f4533bf0b469e6c