[FormEngine] Deprecate FormEngine->disableRTE, remove from QuickEdit
Back in 4.0, there was an option at the bottom of FormEngine to
disable the RTE locally. This was moved to user settings later and
the checkbox removed. The option however is still shown in
Page->QuickEdit. The patch now removes this together with the
code handling this local flag. The user setting option remains.

Deprecates DataPreprocessor->disableRTE
Deprecates FormEngine->disableRTE

Change-Id: Ia939581dad34fea22f9dd4b2e9cd81a631711ae5