[FormEngine] Simplify getListedFields()
getListedFields() has some weird preg_split that hacks the
list of fields into a single field and a palette field depending
on [ and ] marks. This code dates back to Kapsers initial commit
and is never used - no palette is ever rendered, the method is
always filled with a plain list of comma separated fields. The
split code is dropped now together with the palette call.

Change-Id: Id0e808ac7054affab4e2cced14b7e286379101ab