[TASK] Update static composer class loading
The committed composer class loading information has not been
updated for a while. Do so, to add classes that have been added
meanwhile to required TYPO3 extensions to the map and update
class loader code to be current.

Also adapt the .gitignore file to be correct and add a script
to automate this process (so that it can theoretically be done
before a release)

Resolves: #70465
Releases: 6.2
Change-Id: Ife3a9b47a6fbd93e4d10c6e693b597f42a338bc9