[TASK] Create functional test databases with UTF8 character set/collation
Change the create database statements in the functional test setup to
ensure an UTF-8 character set and collation. Otherwise the tests use the
default collation of the server which could lead to database being
created with Non-UTF-8 character sets like latin1_swedish_ci which is
widely in use as a default character set.

Resolves: #67474
Releases: master
Change-Id: Iefdb686a7ceb23e7934702904d16e7d44ae88839