[TASK] Acceptance tests in controlled environment
To successfully run acceptance tests we need to create
a "controlled environment" that is separated from any
existing core environment. A standalone instance is
created that has its own set of loaded extensions,
an own database and an own web path.

The patch refactors and simplifies the existing set up
code of unit and functional test environments to be
better reusable to achieve this.

Change-Id: I6c4a818fa82d2723075888ad8a60b30acb5ed30e
Resolves: #68113
Releases: master