[!!!][FEATURE] Introduce Doctrine DBAL database connections
The Doctrine DBAL library is added as a composer dependency
as a foundation to replace the current DatabaseConnection class,
EXT:dbal and EXT:adodb.

Doctrine DBAL is encapsulated within a ConnectionPool class that
manages the connections to all defined database connections.

The main parts of the patch consist of the connection management,
convenience methods for simple SQL queries, a QueryBuilder to build
complex queries in a database platform independent way and a
QueryRestriction Builder that aims to replace deleteClause and
BEenableFields in the backend context as well as enableFields in
frontend context.

Documentation and an example implementation of using the API will
follow in separate patches.

Releases: master
Resolves: #75454
Change-Id: I47837d9e77331132807bbb7fb956c359031b4f16