[TASK] travis-ci: Limit number of parallel jobs
The docker environment of travis-ci reports no solid amount of
available CPU's, so parallel forks a massive amount of processes
in parallel.
The patch limits the maximum number of jobs in parallel for
travis to 6. This should lead to full CPU usage (there should
be 2 hardware available) even if the one or the other process
is in i/o wait for mysql or similar, and 6 should still not
hammer the VM too much.
This has a number of advantages, since the overall memory
consumption is lower, the system has to do much less context
switches, we're a nicer neighbor, single tests are quicker
and the test order is better predictable.

Change-Id: I0965afb823d1a7d23fddbed8968c2ff313aa5db9
Resolves: #75764
Releases: master