[BUGFIX] Prevent orphaned tags in Typo3DatabaseBackend
The 7.6 and 6.2 implementation of Typo3DatabaseBackend cache backend has
bugs in two methods (in mysql-non-dbal versions). Those were introduced
by #61814 and fixed in master with #77160.

flushByTag() leaves orphaned tags in tags table - if a row has two tags
and flushByTag() is executed on one tag, the other is left.

collectGargabe() does not find orphaned tags collectGarbage() and does
not delete an expired cache row if it has no tags (fix: left outer join)

The patch migrates the functional tests and fixes the issue.

Change-Id: Ie53f54eceb3e47c21c31e7263a3f855b1cb93660
Resolves: #77204
Releases: 7.6, 6.2