[TASK] Replace setLocale() in tests with phpunit backup functionality
There is a phpunit shipped functionality to reliably fiddle with locale
settings and have the valid settings back afterwards.
This funcionality was not used in core unit tests yet, so it is
introduced now.

In this process, some sanitizing to prevent tests from fail due to
non existent locales on a certain system is added.

Resolves: #78387
Releases: master
Change-Id: Iebd9b9251590283e6d2889db09a62a31a595fdb3